src/requirements.txt
RPi.GPIO
sounddevice
soundfile
numpy